+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 CHANGELOG
 =========
+
++ Added a special flag `IS_EMPTY` which can be passed into `assert_body()` to assert an empty response body.